+++ src/omnibase_spi/protocols/onex/__init__.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from .protocol_compute_node import ProtocolComputeNode
+from .protocol_effect_node import ProtocolEffectNode
 from .protocol_onex_envelope import ProtocolOnexEnvelope
 from .protocol_onex_node import ProtocolOnexNode
 from .protocol_onex_reply import ProtocolOnexReply
@@ -15,8 +17,12 @@ from .protocol_onex_validation import (
     ProtocolOnexValidationResult,
 )
 from .protocol_onex_version_loader import ProtocolToolToolOnexVersionLoader
+from .protocol_orchestrator_node import ProtocolOrchestratorNode
+from .protocol_reducer_node import ProtocolReducerNode
 
 __all__ = [
+    "ProtocolComputeNode",
+    "ProtocolEffectNode",
     "ProtocolOnexEnvelope",
     "ProtocolOnexNode",
     "ProtocolOnexReply",
@@ -27,5 +33,7 @@ __all__ = [
     "ProtocolOnexValidation",
     "ProtocolOnexValidationReport",
     "ProtocolOnexValidationResult",
+    "ProtocolOrchestratorNode",
+    "ProtocolReducerNode",
     "ProtocolToolToolOnexVersionLoader",
 ]